修复编译错误, 使 colcon build 通过 (容器化重构前置)
- package.xml: 删除冗余的 exec_depend tf2_ros (与 depend 重复, catkin_pkg 拒绝解析)
- vision/color_utils.hpp: 删除不存在的头文件 opencv2/core/parallel.hpp
  (OpenCV 4.5.4 无此文件; parallel_for_/Range 由 core.hpp 已带入)
- vision/color_utils.cpp: FileNodeIterator 无 name() 成员, 改为 (*it).name()
- camera/building_spot_detector: validate_angles / validate_opposite_sides 去掉
  static (函数体读取 min_angle_score_ / min_side_score_ 成员变量);
  as_int() 返回 int64_t, 与 int 字面量在 std::min/max 推导失败, 补 static_cast<int>
- camera/pillar_detector: 同上两处; 另 PillarImageResult 是命名空间层结构体,
  去掉 PillarDetector:: 限定; std::min(float, double) 补 static_cast<float>
- camera/mustika_detector: MustikaImageResult 同理, 去掉 MustikaDetector:: 限定

+++ src/camera/building_spot_detector.cpp
@@ -682,11 +682,11 @@ rcl_interfaces::msg::SetParametersResult BuildingSpotDetector::on_parameter_chan
   for (const auto& param : params) {
     const std::string& name = param.get_name();
     try {
-           if (name == "morph_close_kernel")     morph_close_kernel_ = std::max(3, std::min(param.as_int(), 31));
-      else if (name == "morph_close_iters")      morph_close_iters_ = std::max(0, std::min(param.as_int(), 5));
+           if (name == "morph_close_kernel")     morph_close_kernel_ = std::max(3, std::min(static_cast<int>(param.as_int()), 31));
+      else if (name == "morph_close_iters")      morph_close_iters_ = std::max(0, std::min(static_cast<int>(param.as_int()), 5));
       else if (name == "approx_epsilon_factor")  approx_epsilon_factor_ = std::max(0.01, std::min(param.as_double(), 0.10));
-      else if (name == "contour_min_area")       contour_min_area_ = std::max(50, param.as_int());
-      else if (name == "contour_max_area")       contour_max_area_ = std::max(contour_min_area_ + 1, param.as_int());
+      else if (name == "contour_min_area")       contour_min_area_ = std::max(50, static_cast<int>(param.as_int()));
+      else if (name == "contour_max_area")       contour_max_area_ = std::max(contour_min_area_ + 1, static_cast<int>(param.as_int()));
       else if (name == "min_green_ratio")        min_green_ratio_ = std::max(0.2, std::min(param.as_double(), 1.0));
       else if (name == "min_convexity")          min_convexity_ = std::max(0.5, std::min(param.as_double(), 1.0));
       else if (name == "min_angle_score")        min_angle_score_ = std::max(0.3, std::min(param.as_double(), 1.0));
